refactor: move diff related code from evo/simplifiedmns to new module evo/smldiff
This commit is to prevent multiple circular dependencies over llmq/ module because
evo/simplifiedmns depends on llmq/blockprocessor, llmq/commitment, llmq/quorums, but they
includes evo/deterministicmns.

Once evo/deterministicmns knows about simplifiedmns it causes 10 more loops over llmq/* and evo/deterministicmns
