+++ src/evo/simplifiedmns.h
@@ -12,18 +12,12 @@
 #include <merkleblock.h>
 #include <netaddress.h>
 #include <pubkey.h>
-#include <sync.h>
-#include <threadsafety.h>
 #include <util/pointer.h>
 
 class UniValue;
-class CBlockIndex;
 class CDeterministicMN;
 class CDeterministicMNList;
-class CDeterministicMNManager;
 class ChainstateManager;
-
-extern RecursiveMutex cs_main;
 
 namespace llmq {
 class CFinalCommitment;
@@ -115,71 +109,6 @@ public:
     uint256 CalcMerkleRoot(bool* pmutated = nullptr) const;
     bool operator==(const CSimplifiedMNList& rhs) const;
 };
-
-/// P2P messages
-
-class CGetSimplifiedMNListDiff
-{
-public:
-    uint256 baseBlockHash;
-    uint256 blockHash;
-
-    SERIALIZE_METHODS(CGetSimplifiedMNListDiff, obj)
-    {
-        READWRITE(obj.baseBlockHash, obj.blockHash);
-    }
-};
-
-class CSimplifiedMNListDiff
-{
-public:
-    static constexpr uint16_t CURRENT_VERSION = 1;
-
-    uint256 baseBlockHash;
-    uint256 blockHash;
-    CPartialMerkleTree cbTxMerkleTree;
-    CTransactionRef cbTx;
-    std::vector<uint256> deletedMNs;
-    std::vector<CSimplifiedMNListEntry> mnList;
-    uint16_t nVersion{CURRENT_VERSION};
-
-    std::vector<std::pair<uint8_t, uint256>> deletedQuorums; // p<LLMQType, quorumHash>
-    std::vector<llmq::CFinalCommitment> newQuorums;
-
-    // Map of Chainlock Signature used for shuffling per set of quorums
-    // The set of quorums is the set of indexes corresponding to entries in newQuorums
-    std::map<CBLSSignature, std::set<uint16_t>> quorumsCLSigs;
-
-    SERIALIZE_METHODS(CSimplifiedMNListDiff, obj)
-    {
-        if ((s.GetType() & SER_NETWORK) && s.GetVersion() >= MNLISTDIFF_VERSION_ORDER) {
-            READWRITE(obj.nVersion);
-        }
-        READWRITE(obj.baseBlockHash, obj.blockHash, obj.cbTxMerkleTree, obj.cbTx);
-        if ((s.GetType() & SER_NETWORK) && s.GetVersion() >= BLS_SCHEME_PROTO_VERSION && s.GetVersion() < MNLISTDIFF_VERSION_ORDER) {
-            READWRITE(obj.nVersion);
-        }
-        READWRITE(obj.deletedMNs, obj.mnList);
-        READWRITE(obj.deletedQuorums, obj.newQuorums);
-        if ((s.GetType() & SER_NETWORK) && s.GetVersion() >= MNLISTDIFF_CHAINLOCKS_PROTO_VERSION) {
-            READWRITE(obj.quorumsCLSigs);
-        }
-    }
-
-    CSimplifiedMNListDiff();
-    ~CSimplifiedMNListDiff();
-
-    bool BuildQuorumsDiff(const CBlockIndex* baseBlockIndex, const CBlockIndex* blockIndex,
-                          const llmq::CQuorumBlockProcessor& quorum_block_processor);
-    bool BuildQuorumChainlockInfo(const llmq::CQuorumManager& qman, const CBlockIndex* blockIndex);
-
-    [[nodiscard]] UniValue ToJson(bool extended = false) const;
-};
-
-bool BuildSimplifiedMNListDiff(CDeterministicMNManager& dmnman, const ChainstateManager& chainman,
-                               const llmq::CQuorumBlockProcessor& qblockman, const llmq::CQuorumManager& qman,
-                               const uint256& baseBlockHash, const uint256& blockHash, CSimplifiedMNListDiff& mnListDiffRet,
-                               std::string& errorRet, bool extended = false) EXCLUSIVE_LOCKS_REQUIRED(::cs_main);
 
 bool CalcCbTxMerkleRootMNList(uint256& merkleRootRet, CSimplifiedMNList&& sml, BlockValidationState& state);
